Configurable active detection prior
Allow users to provide a prior for the active detection probability in
the form of a beta distribution. Also adjusted the model away from a
logit-normal prior distribution for phi to a beta prior distribution.
Default argument does not quite align with the previous default, but is
similarly informative in the below ~30% region.