Add `SeverityEstimateModel` class
Added a new S4 class to represent a model that will contain the metadata
needed for building a model as well as the `SeverityEstimateModel` 
constructor function for creating an instance of this object. Concept is
that this will be the first line in a model definition and then the 
returned value can be piped along.

Also added a `print` method for `SeverityEstimateModel`. Currently just
a thin wrapper around `str`, but want a place holder for the future when
the object structure is better defined.

Closes #47.