Add checks for var/sd in beta params
Added a check to the `beta_parameterization` function for bounds on the
variance, 'var' or 'sd', with respect to the given mean when using the
'mean'/'var' or 'mean'/'sd' parameterizations. Without this check the
`beta_parameterization` function could return negative values for alpha
and beta without error, i.e if given `c("mean" = 0.5, "var" = 0.26)`.