Check built package, refactor `Makefile`
Instead of running check on the directory the `check` target now checks
the built package. Also added a `build` target that the `check` target
can depend on. Minor refactoring to the `Makefile` to make the `.PHONY`
targets more maintainable.

Added a pandoc install step to the `setup-r` custom action, required by
`R CMD build`. Added '.jj' directory to the `.Rbuildignore`.